Update to the last Rust.
Verson of rustc: 0.10-pre (9b9e2f8 2014-02-22 20:21:54 -0800)

+++ src/util/hash_map.rs
@@ -157,7 +157,7 @@ impl<K: Eq, V, H: HashFun<K>> HashMap<K, V, H> {
             self.table.push(Entry::new(key, value));
             self.num_elem = self.num_elem + 1;
 
-            (true, (self.num_elem - 1) as uint)
+            (true, self.num_elem - 1)
         }
         else {
             if replace {
